Completed Post API (refactor)
