+++ src/CodeSharing.Server/Services/Interfaces/ILabelService.cs
@@ -1,4 +1,6 @@
+using CodeSharing.Server.Datas.Entities;
 using CodeSharing.ViewModels.Contents.Label;
+using CodeSharing.ViewModels.Contents.Post;
 
 namespace CodeSharing.Server.Services.Interfaces;
