+++ src/CodeSharing.Utilities/Helpers/FunctionBase.cs
@@ -35,4 +35,6 @@ public static class FunctionBase
     {
         return context + "_" + DateTime.Now.ToString("yyyyMMddHHmmss") + "_" + Guid.NewGuid().ToString("N");
     }
+    
+    
 }
